requirements.txt
gidgethub
requests
cachetools
aiohttp
setuptools_scm
pytest
pytest-cov
pytest-xdist
pip
GitPython
pre-commit
gcg